Move IOrigin into a separate file
